+++ src/components/ToastProvider/ToastProvider.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import useKeydown from '../../hooks/use-keydown';
+
 export const ToastContext = React.createContext();
 
 function ToastProvider({ children }) {
@@ -16,19 +18,11 @@ function ToastProvider({ children }) {
     },
   ]);
 
-  React.useEffect(() => {
-    function handleKeyDown(event) {
-      if (event.code === 'Escape') {
-        setToasts([]);
-      }
-    }
-
-    window.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-    };
+  const handleEscape = React.useCallback(() => {
+    setToasts([]);
   }, []);
+
+  useKeydown('Escape', handleEscape);
 
   function createToast(message, variant) {
     const nextToasts = [
